requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
python-dateutil==2.9.0.post0
icalendar==5.0.12
pyyaml==6.0.2
